Clean up favorite button
